Add support for SHA-2.

+++ src/cert.c
@@ -349,7 +349,7 @@ ksba_cert_get_image (ksba_cert_t cert, size_t *r_length )
     {
       fprintf (stderr,"\nOops, ksba_cert_get_image failed: "
                "imagelen=%d  hdr=%d len=%d off=%d\n",
-               cert->imagelen, n->nhdr, n->len, n->off);
+               cert->imagelen, n->nhdr, (int)n->len, n->off);
       return NULL;
     }
   /*assert (n->nhdr + n->len + n->off <= cert->imagelen);*/
